refactor(order-products): extract shared OrderProducts component and groupProductItemsByShipmentId @W-23444948
